% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/def.R
\name{append_make_def}
\alias{append_make_def}
\title{Appends a Makefile variable definition to a Makefile}
\usage{
append_make_def(makefile, variable, definition)
}
\arguments{
\item{makefile}{A Makefile created by \code{\link{makefile}}}

\item{variable}{Variable name}

\item{definition}{Definition for this variable}
}
\value{
The first parameter, with the newly created rule appended
}
\description{
This helper function creates a variable definition and appends it to an
existing Makefile.
Most useful in pipes.
}
\examples{
library(magrittr)
makefile() \%>\%
  append_make_def("R_USER_LIBRARY", .libPaths()[[1L]])
}
\seealso{
\code{\link{make_def}}, \code{\link{makefile}}
}

