% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rule.R
\name{append_make_rule}
\alias{append_make_rule}
\title{Appends a Makefile rule to a Makefile}
\usage{
append_make_rule(makefile, targets, deps = NULL, script = NULL)
}
\arguments{
\item{makefile}{A Makefile created by \code{\link{makefile}}}

\item{targets}{Target names as a character vector}

\item{deps}{Dependency names as a character vector}

\item{script}{A script to execute to build the targets.}
}
\value{
The first parameter, with the newly created rule appended
}
\description{
This helper function creates a rule and appends it to an existing Makefile.
Most useful in pipes.
}
\examples{
library(magrittr)
makefile() \%>\%
  append_make_rule("all", c("first_target", "second_target")) \%>\%
  append_make_rule(".FORCE") \%>\%
  append_make_rule("first_target", ".FORCE", "echo 'Building first target'") \%>\%
  append_make_rule("second_target", "first_target",
    c("echo 'Building second target'", "echo 'Done'"))
}
\seealso{
\code{\link{make_rule}}, \code{\link{create_makefile}}
}

