% Generated by roxygen2 (4.1.1.9001): do not edit by hand
% Please edit documentation in R/group.R
\name{c.MakefileR_group}
\alias{+.MakefileR_group}
\alias{c.MakefileR_group}
\title{Concatenation of rules}
\usage{
\method{c}{MakefileR_group}(..., recursive = FALSE)

\method{+}{MakefileR_group}(x, y)
}
\arguments{
\item{...,x,y}{Objects of class \code{MakefileR}, the first
(\code{x} or the first element of \code{...})
must be of class \code{MakefileR_group}
(created by \code{\link{make_group}} or \code{\link{makefile}})}

\item{recursive}{unused}
}
\description{
Rules can be appended to groups and Makefiles using the
\code{\link[base]{c}} function or the \code{\link[base]{+}} operator.
}
\examples{
c(make_group(sep = ""),
  make_group(make_comment("Dummy targets"),
             make_rule(".FORCE"), make_rule(".SILENT")),
  make_group(make_comment("Definitions"),
             make_def("A", "a")))

makefile() + (make_group() + make_comment("Definitions") + make_def("A", "a"))
}

