% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/def.R
\name{create_make_def}
\alias{create_make_def}
\title{Creates a variable definition in a Makefile}
\usage{
create_make_def(variable, definition)
}
\arguments{
\item{variable}{Variable name}

\item{definition}{Definition for this variable}
}
\description{
A variable definition in a \code{Makefile} consists of a variable name
and its defition.  Both are separated by the equality sign \code{=}.
}
\details{
No quoting is applied to the definition by this function.
Currently, both variable and definition are required to be character values
of length one.
}
\examples{
create_make_def("R_USER_LIBRARY", .libPaths()[[1L]])
}
\references{
\url{https://www.gnu.org/software/make/manual/}
}

