% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comment.R
\name{make_comment}
\alias{make_comment}
\title{Creates a Makefile comment}
\usage{
make_comment(...)
}
\arguments{
\item{...}{Character vector(s) without leading hash \code{#}}
}
\description{
For helping the reader understand what's happening
}
\examples{
make_comment("This is a comment")))
}
\references{
\url{https://www.gnu.org/software/make/manual/}
}

