% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{make_text}
\alias{make_text}
\title{Creates a custom Makefile entry}
\usage{
make_text(...)
}
\arguments{
\item{...}{\code{[character]}\cr Custom text}
}
\value{
An object of class \code{MakefileR_text}
}
\description{
For anything else not covered by the other \code{make_} functions, such as
the \code{export} statement for exporting Makefile variables.
}
\details{
Use the
\code{\link[base]{c}} function or the \code{\link[base]{+}} operator
to append comments to groups and Makefiles.
}
\examples{
make_text("export SOME_VARIABLE")

}
\references{
\url{https://www.gnu.org/software/make/manual/}
}
\seealso{
Other items: \code{\link{make_comment}},
  \code{\link{make_def}}, \code{\link{make_group}},
  \code{\link{make_rule}}
}

