% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/file.R
\name{makefile}
\alias{makefile}
\title{Creates a Makefile}
\usage{
makefile(..., .dots = NULL)
}
\arguments{
\item{...}{Items created by \code{\link{make_rule}} or other \code{make_}
functions}

\item{.dots}{A list rules in addition to \code{...}}
}
\value{
An object of class \code{MakefileR_file}
}
\description{
A \code{Makefile} consists of a list of rules.
}
\details{
Use the
\code{\link[base]{c}} function or the \code{\link[base]{+}} operator
to append rules, definitions, comments and groups.
}
\examples{
makefile(make_rule("all", c("first_target", "second_target")))
}
\references{
\url{https://www.gnu.org/software/make/manual/}
}
\seealso{
\code{\link{make_rule}}, \code{\link{make_def}},
  \code{\link{make_group}}, \code{\link{make_comment}},
  \code{\link{c.MakefileR_group}}
}

