% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_makefile}
\alias{write_makefile}
\title{Writes a Makefile to a file}
\usage{
write_makefile(makefile, file_name)
}
\arguments{
\item{makefile}{\code{[MakefileR]}\cr A Makefile, created by
\code{\link{makefile}}}

\item{file_name}{\code{[character(1)]}\cr Target file name}
}
\value{
The value returned by \code{\link{writeLines}}
}
\description{
Makefiles, as created by \code{\link{makefile}}, only exist in memory until
they are written to a file by this function.
}

