% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/file.R
\name{create_makefile}
\alias{create_makefile}
\title{Creates a Makefile}
\usage{
create_makefile(..., .dots = NULL)
}
\arguments{
\item{...}{Rules created by \code{\link{create_make_rule}}}

\item{.dots}{A list rules in addition to \code{...}}
}
\description{
A \code{Makefile} consists of a list of rules.
}
\examples{
create_makefile(create_make_rule("all", c("first_target", "second_target")))
}
\references{
\url{https://www.gnu.org/software/make/manual/}
}

