% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/group.R
\name{make_group}
\alias{make_group}
\title{Creates a group of Makefile items}
\usage{
make_group(..., .dots = NULL, sep = NULL)
}
\arguments{
\item{...}{Rules created by \code{\link{make_rule}}}

\item{.dots}{A list rules in addition to \code{...}}

\item{sep}{Separator between group items, \code{NULL} (the default) means
  no separator.}
}
\description{
Helps separating similar rules.
}
\examples{
makefile(make_rule("all", c("first_target", "second_target")))
}
\references{
\url{https://www.gnu.org/software/make/manual/}
}

