% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/file.R
\name{makefile}
\alias{makefile}
\title{Creates a Makefile}
\usage{
makefile(..., .dots = NULL)
}
\arguments{
\item{...}{Rules created by \code{\link{make_rule}}}

\item{.dots}{A list rules in addition to \code{...}}
}
\value{
An object of class \code{MakefileR_file}
}
\description{
A \code{Makefile} consists of a list of rules.
}
\examples{
makefile(make_rule("all", c("first_target", "second_target")))
}
\references{
\url{https://www.gnu.org/software/make/manual/}
}

